+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+-   Show progress when checking WindowsPerf installation.
+
 ## 1.2.0
 
 -   Changed "Run System Check" buttons and "Check Installation" command name to "Check WindowsPerf Installation".